+++ Sources/Plasma/NucleusLib/inc/CMakeLists.txt
@@ -16,6 +16,7 @@ set(pnNucleusInc_HEADERS
     plgDispatch.h
     plLoggable.h
     plPhysical.h
+    plPipeDebugFlags.h
     plPipeline.h
     plPipeResReq.h
     plProfile.h
