+++ Sources/Plasma/PubUtilLib/plDrawable/plDrawableSpans.cpp
@@ -63,6 +63,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "hsStream.h"
 #include "hsResMgr.h"
 #include "plPipeline.h"
+#include "plPipeDebugFlags.h"
 #include "plGeometrySpan.h"
 #include "plSpaceTree.h"
 #include "plParticleFiller.h"
@@ -78,7 +79,6 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plSurface/plLayerInterface.h"
 #include "plPipeline/plFogEnvironment.h"
 #include "plPipeline/hsGDeviceRef.h"
-#include "plPipeline/plPipeDebugFlags.h"
 #include "pnMessage/plRefMsg.h"
 #include "pnMessage/plNodeRefMsg.h" 
 #include "pnMessage/plDISpansMsg.h"
