+++ Sources/Plasma/PubUtilLib/plGLight/plLightInfo.cpp
@@ -65,7 +65,7 @@ You can contact Cyan Worlds, Inc. by email legal@cyan.com
 #include "plScene/plRenderRequest.h"
 #include "plPipeline.h"
 #include "plIntersect/plSoftVolume.h"
-#include "plPipeline/plPipeDebugFlags.h"
+#include "plPipeDebugFlags.h"
 #include "pnMessage/plPipeResMakeMsg.h"
 
 #include "plScene/plVisRegion.h"
